[3227] Move AllocationsController into providers directory
to address peer review comments
