+++ app/helpers/breadcrumb_helper.rb
@@ -71,7 +71,7 @@ module BreadcrumbHelper
   end
 
   def allocations_breadcrumb
-    path = allocations_provider_recruitment_cycle_path(@provider.provider_code, @provider.recruitment_cycle_year)
+    path = provider_recruitment_cycle_allocations_path(@provider.provider_code, @provider.recruitment_cycle_year)
     provider_breadcrumb << ["Request PE courses for 2021/22", path]
   end
 end
