+++ app/views/recruitment_cycles/_allocations.html.erb
@@ -5,7 +5,7 @@
     </div>
     <h2 class="govuk-heading-m">
       <%= link_to "Request PE courses for 2021/22",
-                  allocations_provider_recruitment_cycle_path(@provider.provider_code, year),
+                  provider_recruitment_cycle_allocations_path(@provider.provider_code, year),
                   class: "govuk-link",
                   data: { qa: 'request_allocations_link' } %>
     </h2>
